sipXivr: Add WebServer for RESTful interface to Voicemail
Adds a Jetty based Webserver.  Moves the existing mwistatus servlet
into that on path "/mwi", and adds a rudimentary mailbox servlet as well
on path "/mailbox/*"

Use sipXconfig's RESTful interface to enable/disable Auto Attendant's
   special mode.  (XX-6632)

Add support for marking messages heard/unheard via RESTful interface:
   /mailbox/{mailbox}/message/{message id}/heard 
      GET returns status <heard>true|false</heard>
      DELETE marks message unheard and updates MWI
      PUT (no data) marks message heard and updates MWI

Add support for updating a mailbox's MWI
   /mailbox/{mailbox}/mwi
      GET returns the message summary content (in RFC-3842 format)
      PUT (no data) updates the MWI for this mailbox (i.e. tells the status 
         server to send the MWI status to subscriber)

Fixes:
   XX-6632 - Enabling and Disabling options for Special auto attendant 
      from TUI is not working as expected.

Part of:
   XX-6527 - sipXconfig does not mark messages as read with new voicemail 
      system

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16648 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXivr/src/main/java/org/sipfoundry/voicemail/VmMessage.java
@@ -11,20 +11,11 @@ package org.sipfoundry.voicemail;
 import java.io.File;
 import java.io.IOException;
 import java.io.SequenceInputStream;
-import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javax.mail.MessagingException;
-import javax.mail.Multipart;
-import javax.mail.Part;
-import javax.mail.Session;
-import javax.mail.Transport;
 import javax.mail.Flags.Flag;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeBodyPart;
-import javax.mail.internet.MimeMessage;
-import javax.mail.internet.MimeMultipart;
 import javax.sound.sampled.AudioFileFormat;
 import javax.sound.sampled.AudioInputStream;
 import javax.sound.sampled.AudioSystem;
@@ -32,7 +23,6 @@ import javax.sound.sampled.AudioSystem;
 import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
 import org.sipfoundry.sipxivr.Mailbox;
-import org.sipfoundry.sipxivr.MailboxPreferences;
 import org.sipfoundry.voicemail.MessageDescriptor.Priority;
 
 import com.sun.mail.imap.IMAPMessage;
@@ -127,8 +117,8 @@ public class VmMessage {
                  
         // Check for -00.sta
         File status = new File(directory, id+"-00.sta");
+        me.m_statusFile = status;
         if (status.exists()) {
-            me.m_statusFile = status;
             me.m_unHeard = true ;
         } 
 
@@ -429,7 +419,9 @@ public class VmMessage {
      */
     private static File moveFileToDirectory(File file, File newDirectory) throws IOException {
         if (file != null) {
-            FileUtils.moveFileToDirectory(file, newDirectory, false);
+            if (file.exists()) {
+                FileUtils.moveFileToDirectory(file, newDirectory, false);
+            }
             return new File(newDirectory, file.getName());
         }
         return null;
@@ -543,10 +535,25 @@ public class VmMessage {
         if (m_unHeard) {
             if (m_statusFile != null) {
                 FileUtils.deleteQuietly(m_statusFile);
-                m_statusFile = null;
             }
             LOG.info(String.format("VmMessage::markHeard %s marked heard", m_descriptorFile.getPath()));
             m_unHeard = false;
+        }
+    }
+
+    public void markUnheard() {
+        if (!m_unHeard) {
+            if (m_statusFile != null) {
+                String operation= "update status";
+                try {
+                    operation = "creating status file "+m_statusFile.getPath();
+                    FileUtils.touch(m_statusFile);
+                    LOG.info(String.format("VmMessage::markHeard %s marked unheard", m_descriptorFile.getPath()));
+                    m_unHeard = false;
+                } catch (IOException e) {
+                    LOG.error("VmMessage::markUnHeard error while "+operation, e);
+                }
+            } 
         }
     }
